requirements.txt
airflow-provider-great-expectations==0.1.4
alembic==1.7.7
altair==4.2.0
anyio==3.6.1
apache-airflow==2.3.1
apache-airflow-providers-apache-spark==2.1.3
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apache-sedona==1.2.0
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==20.3.0
Babel==2.10.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachelib==0.7.0
cattrs==1.10.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-plugins==1.1.1
clickclick==20.10.2
cligj==0.7.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Fiona==1.8.21
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
graphviz==0.20
great-expectations==0.15.7
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
inflection==0.5.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==2.4.7
pyproj==3.3.1
pyrsistent==0.18.1
pyspark==3.2.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.0.0
requests==2.27.1
rfc3986==1.5.0
rich==12.4.4
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.8.1
Send2Trash==1.8.0
setproctitle==1.2.3
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
stack-data==0.2.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
text-unidecode==1.3
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
unicodecsv==0.14.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
